+++ src/new-components/divider/divider.tsx
@@ -6,10 +6,10 @@ import type { DividerWeight } from '../common-types'
 import styles from './divider.module.css'
 
 interface DividerProps {
-    weight: Exclude<DividerWeight, 'none'>
+    weight?: Exclude<DividerWeight, 'none'>
 }
 
-function Divider({ weight, ...props }: DividerProps) {
+function Divider({ weight = 'tertiary', ...props }: DividerProps) {
     return <Box as="hr" className={getClassNames(styles, 'weight', weight)} {...props} />
 }